Renaming package to fuse_models
