+++ fuse_models/CHANGELOG.rst
@@ -1,6 +1,12 @@
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-Changelog for package fuse_rl
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+Changelog for package fuse_models
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+0.4.0 (2019-08-14)
+------------------
+
+0.3.0 (2019-08-14)
+------------------
 
 0.2.0 (2019-07-12)
 ------------------
