+++ fuse_models/include/fuse_models/parameters/odometry_2d_model_params.h
@@ -31,10 +31,10 @@
  *  ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  *  POSSIBILITY OF SUCH DAMAGE.
  */
-#ifndef FUSE_RL_PARAMETERS_ODOMETRY_2D_MODEL_PARAMS_H
-#define FUSE_RL_PARAMETERS_ODOMETRY_2D_MODEL_PARAMS_H
+#ifndef FUSE_MODELS_PARAMETERS_ODOMETRY_2D_MODEL_PARAMS_H
+#define FUSE_MODELS_PARAMETERS_ODOMETRY_2D_MODEL_PARAMS_H
 
-#include <fuse_rl/parameters/parameter_base.h>
+#include <fuse_models/parameters/parameter_base.h>
 
 #include <fuse_variables/orientation_2d_stamped.h>
 #include <fuse_variables/position_2d_stamped.h>
@@ -46,7 +46,7 @@
 #include <vector>
 
 
-namespace fuse_rl
+namespace fuse_models
 {
 
 namespace parameters
@@ -92,6 +92,6 @@ struct Odometry2DModelParams : public ParameterBase
 
 }  // namespace parameters
 
-}  // namespace fuse_rl
+}  // namespace fuse_models
 
-#endif  // FUSE_RL_PARAMETERS_ODOMETRY_2D_MODEL_PARAMS_H
+#endif  // FUSE_MODELS_PARAMETERS_ODOMETRY_2D_MODEL_PARAMS_H
