+++ fuse_models/include/fuse_models/parameters/unicycle_2d_ignition_params.h
@@ -31,10 +31,10 @@
  *  ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  *  POSSIBILITY OF SUCH DAMAGE.
  */
-#ifndef FUSE_RL_PARAMETERS_UNICYCLE_2D_IGNITION_PARAMS_H
-#define FUSE_RL_PARAMETERS_UNICYCLE_2D_IGNITION_PARAMS_H
+#ifndef FUSE_MODELS_PARAMETERS_UNICYCLE_2D_IGNITION_PARAMS_H
+#define FUSE_MODELS_PARAMETERS_UNICYCLE_2D_IGNITION_PARAMS_H
 
-#include <fuse_rl/parameters/parameter_base.h>
+#include <fuse_models/parameters/parameter_base.h>
 
 #include <ros/node_handle.h>
 
@@ -44,7 +44,7 @@
 #include <vector>
 
 
-namespace fuse_rl
+namespace fuse_models
 {
 
 namespace parameters
@@ -160,6 +160,6 @@ struct Unicycle2DIgnitionParams : public ParameterBase
 
 }  // namespace parameters
 
-}  // namespace fuse_rl
+}  // namespace fuse_models
 
-#endif  // FUSE_RL_PARAMETERS_UNICYCLE_2D_IGNITION_PARAMS_H
+#endif  // FUSE_MODELS_PARAMETERS_UNICYCLE_2D_IGNITION_PARAMS_H
