+++ fuse_models/include/fuse_models/pose_2d/model.h
@@ -31,47 +31,50 @@
  *  ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  *  POSSIBILITY OF SUCH DAMAGE.
  */
-#ifndef FUSE_RL_ACCELERATION_2D_MODEL_H
-#define FUSE_RL_ACCELERATION_2D_MODEL_H
+#ifndef FUSE_MODELS_POSE_2D_MODEL_H
+#define FUSE_MODELS_POSE_2D_MODEL_H
 
-#include <fuse_rl/parameters/acceleration_2d_model_params.h>
+#include <fuse_models/parameters/pose_2d_model_params.h>
 
 #include <fuse_core/async_sensor_model.h>
 #include <fuse_core/uuid.h>
 
-#include <geometry_msgs/AccelWithCovarianceStamped.h>
+#include <geometry_msgs/PoseWithCovarianceStamped.h>
 #include <ros/ros.h>
 #include <tf2_ros/transform_listener.h>
 
 
-namespace fuse_rl
+namespace fuse_models
 {
 
-namespace acceleration_2d
+namespace pose_2d
 {
 
 /**
- * @brief An adapter-type sensor that produces 2D linear acceleration constraints from information published by another
- *   node
+ * @brief An adapter-type sensor that produces absolute or relative pose constraints from information published by
+ * another node.
  *
- * This sensor subscribes to a geometry_msgs::AccelWithCovarianceStamped topic and converts each received message
- * into an 2D linear acceleration variable and constraint.
+ * This sensor subscribes to a geometry_msgs::PoseWithCovarianceStamped topic and converts each received message
+ * into an absolute or relative pose constraint. If the \p differential parameter is set to false (the default), the
+ * measurement will be treated as an absolute constraint. If it is set to true, consecutive measurements will be used
+ * to generate relative pose constraints.
  *
  * Parameters:
  *  - device_id (uuid string, default: 00000000-0000-0000-0000-000000000000) The device/robot ID to publish
  *  - device_name (string) Used to generate the device/robot ID if the device_id is not provided
- *  - queue_size (int, default: 10) The subscriber queue size for the twist messages
- *  - target_frame (string) The target frame_id to transform the data into before using it
- *  - topic (string) The topic to which to subscribe for the twist messages
+ *  - queue_size (int, default: 10) The subscriber queue size for the pose messages
+ *  - topic (string) The topic to which to subscribe for the pose messages (required if \p subscribe is true)
+ *  - differential (bool, default: false) Whether we should fuse measurements absolutely, or to create relative pose
+ *      constraints using consecutive measurements.
  *
  * Subscribes:
- *  - \p topic (geometry_msgs::AccelWithCovarianceStamped) Acceleration information at a given timestamp
+ *  - \p topic (geometry_msgs::PoseWithCovarianceStamped) Absolute pose information at a given timestamp
  */
 class Model : public fuse_core::AsyncSensorModel
 {
 public:
   SMART_PTR_DEFINITIONS(Model);
-  using ParameterType = parameters::Acceleration2DModelParams;
+  using ParameterType = parameters::Pose2DModelParams;
 
   /**
    * @brief Default constructor
@@ -84,10 +87,10 @@ public:
   virtual ~Model() = default;
 
   /**
-   * @brief Callback for acceleration messages
-   * @param[in] msg - The acceleration message to process
+   * @brief Callback for pose messages
+   * @param[in] msg - The pose message to process
    */
-  void process(const geometry_msgs::AccelWithCovarianceStamped::ConstPtr& msg);
+  void process(const geometry_msgs::PoseWithCovarianceStamped::ConstPtr& msg);
 
 protected:
   fuse_core::UUID device_id_;  //!< The UUID of this device
@@ -113,6 +116,8 @@ protected:
 
   ParameterType params_;
 
+  geometry_msgs::PoseWithCovarianceStamped::ConstPtr previous_pose_msg_;
+
   tf2_ros::Buffer tf_buffer_;
 
   tf2_ros::TransformListener tf_listener_;
@@ -120,8 +125,8 @@ protected:
   ros::Subscriber subscriber_;
 };
 
-}  // namespace acceleration_2d
+}  // namespace pose_2d
 
-}  // namespace fuse_rl
+}  // namespace fuse_models
 
-#endif  // FUSE_RL_ACCELERATION_2D_MODEL_H
+#endif  // FUSE_MODELS_POSE_2D_MODEL_H
