+++ fuse_models/src/acceleration_2d/model.cpp
@@ -31,8 +31,8 @@
  *  ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  *  POSSIBILITY OF SUCH DAMAGE.
  */
-#include <fuse_rl/common/sensor_proc.h>
-#include <fuse_rl/acceleration_2d/model.h>
+#include <fuse_models/common/sensor_proc.h>
+#include <fuse_models/acceleration_2d/model.h>
 
 #include <fuse_core/transaction.h>
 #include <fuse_core/uuid.h>
@@ -43,9 +43,9 @@
 
 
 // Register this sensor model with ROS as a plugin.
-PLUGINLIB_EXPORT_CLASS(fuse_rl::acceleration_2d::Model, fuse_core::SensorModel)
+PLUGINLIB_EXPORT_CLASS(fuse_models::acceleration_2d::Model, fuse_core::SensorModel)
 
-namespace fuse_rl
+namespace fuse_models
 {
 
 namespace acceleration_2d
@@ -105,4 +105,4 @@ void Model::process(const geometry_msgs::AccelWithCovarianceStamped::ConstPtr& m
 
 }  // namespace acceleration_2d
 
-}  // namespace fuse_rl
+}  // namespace fuse_models
