+++ fuse_models/src/odometry_2d/model.cpp
@@ -31,8 +31,8 @@
  *  ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  *  POSSIBILITY OF SUCH DAMAGE.
  */
-#include <fuse_rl/common/sensor_proc.h>
-#include <fuse_rl/odometry_2d/model.h>
+#include <fuse_models/common/sensor_proc.h>
+#include <fuse_models/odometry_2d/model.h>
 
 #include <fuse_core/transaction.h>
 #include <fuse_core/uuid.h>
@@ -45,9 +45,9 @@
 
 
 // Register this sensor model with ROS as a plugin.
-PLUGINLIB_EXPORT_CLASS(fuse_rl::odometry_2d::Model, fuse_core::SensorModel)
+PLUGINLIB_EXPORT_CLASS(fuse_models::odometry_2d::Model, fuse_core::SensorModel)
 
-namespace fuse_rl
+namespace fuse_models
 {
 
 namespace odometry_2d
@@ -153,4 +153,4 @@ void Model::process(const nav_msgs::Odometry::ConstPtr& msg)
 
 }  // namespace odometry_2d
 
-}  // namespace fuse_rl
+}  // namespace fuse_models
