+++ fuse_models/test/test_sensor_proc.cpp
@@ -3,14 +3,14 @@
  * Unauthorized copying of this file, via any medium, is strictly prohibited
  * Proprietary and confidential
  ***************************************************************************/
-#include <fuse_rl/common/sensor_proc.h>
+#include <fuse_models/common/sensor_proc.h>
 
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
 
 #include <vector>
 
-namespace fuse_rl_common = fuse_rl::common;
+namespace fm_common = fuse_models::common;
 
 TEST(TestSuite, mergeXYPositionAndYawOrientationIndices)
 {
@@ -19,7 +19,7 @@ TEST(TestSuite, mergeXYPositionAndYawOrientationIndices)
 
   const size_t orientation_offset = 2;
 
-  const auto merged_indices = fuse_rl_common::mergeIndices(position_indices, orientation_indices, orientation_offset);
+  const auto merged_indices = fm_common::mergeIndices(position_indices, orientation_indices, orientation_offset);
 
   EXPECT_EQ(position_indices.size() + orientation_indices.size(), merged_indices.size());
   EXPECT_THAT(position_indices,
@@ -34,7 +34,7 @@ TEST(TestSuite, mergeXPositionAndYawOrientationIndices)
 
   const size_t orientation_offset = 2;
 
-  const auto merged_indices = fuse_rl_common::mergeIndices(position_indices, orientation_indices, orientation_offset);
+  const auto merged_indices = fm_common::mergeIndices(position_indices, orientation_indices, orientation_offset);
 
   EXPECT_EQ(position_indices.size() + orientation_indices.size(), merged_indices.size());
   EXPECT_THAT(position_indices,
@@ -49,7 +49,7 @@ TEST(TestSuite, mergeXYPositionAndEmptyOrientationIndices)
 
   const size_t orientation_offset = 2;
 
-  const auto merged_indices = fuse_rl_common::mergeIndices(position_indices, orientation_indices, orientation_offset);
+  const auto merged_indices = fm_common::mergeIndices(position_indices, orientation_indices, orientation_offset);
 
   EXPECT_EQ(position_indices.size(), merged_indices.size());
   EXPECT_THAT(position_indices, testing::ElementsAreArray(merged_indices));
@@ -62,7 +62,7 @@ TEST(TestSuite, mergeEmptyPositionAndYawOrientationIndices)
 
   const size_t orientation_offset = 2;
 
-  const auto merged_indices = fuse_rl_common::mergeIndices(position_indices, orientation_indices, orientation_offset);
+  const auto merged_indices = fm_common::mergeIndices(position_indices, orientation_indices, orientation_offset);
 
   EXPECT_EQ(orientation_indices.size(), merged_indices.size());
   EXPECT_EQ(orientation_indices.back() + orientation_offset, merged_indices.back());
@@ -75,7 +75,7 @@ TEST(TestSuite, mergeEmptyPositionAndEmptyOrientationIndices)
 
   const size_t orientation_offset = 2;
 
-  const auto merged_indices = fuse_rl_common::mergeIndices(position_indices, orientation_indices, orientation_offset);
+  const auto merged_indices = fm_common::mergeIndices(position_indices, orientation_indices, orientation_offset);
 
   EXPECT_TRUE(merged_indices.empty());
 }
@@ -93,12 +93,12 @@ TEST(TestSuite, populatePartialMeasurementXYPositionYawOrientation)
 
   const size_t orientation_offset = 2;
 
-  const auto merged_indices = fuse_rl_common::mergeIndices(position_indices, orientation_indices, orientation_offset);
+  const auto merged_indices = fm_common::mergeIndices(position_indices, orientation_indices, orientation_offset);
 
   fuse_core::VectorXd pose_mean_partial(position_indices.size() + orientation_indices.size());
   fuse_core::MatrixXd pose_covariance_partial(pose_mean_partial.rows(), pose_mean_partial.rows());
 
-  fuse_rl_common::populatePartialMeasurement(
+  fm_common::populatePartialMeasurement(
     pose_mean,
     pose_covariance,
     merged_indices,
@@ -122,12 +122,12 @@ TEST(TestSuite, populatePartialMeasurementXPositionYawOrientation)
 
   const size_t orientation_offset = 2;
 
-  const auto merged_indices = fuse_rl_common::mergeIndices(position_indices, orientation_indices, orientation_offset);
+  const auto merged_indices = fm_common::mergeIndices(position_indices, orientation_indices, orientation_offset);
 
   fuse_core::VectorXd pose_mean_partial(position_indices.size() + orientation_indices.size());
   fuse_core::MatrixXd pose_covariance_partial(pose_mean_partial.rows(), pose_mean_partial.rows());
 
-  fuse_rl_common::populatePartialMeasurement(
+  fm_common::populatePartialMeasurement(
     pose_mean,
     pose_covariance,
     merged_indices,
@@ -165,12 +165,12 @@ TEST(TestSuite, populatePartialMeasurementEmptyPositionYawOrientation)
 
   const size_t orientation_offset = 2;
 
-  const auto merged_indices = fuse_rl_common::mergeIndices(position_indices, orientation_indices, orientation_offset);
+  const auto merged_indices = fm_common::mergeIndices(position_indices, orientation_indices, orientation_offset);
 
   fuse_core::VectorXd pose_mean_partial(position_indices.size() + orientation_indices.size());
   fuse_core::MatrixXd pose_covariance_partial(pose_mean_partial.rows(), pose_mean_partial.rows());
 
-  fuse_rl_common::populatePartialMeasurement(
+  fm_common::populatePartialMeasurement(
     pose_mean,
     pose_covariance,
     merged_indices,
@@ -194,12 +194,12 @@ TEST(TestSuite, populatePartialMeasurementXYPositionEmptyOrientation)
 
   const size_t orientation_offset = 2;
 
-  const auto merged_indices = fuse_rl_common::mergeIndices(position_indices, orientation_indices, orientation_offset);
+  const auto merged_indices = fm_common::mergeIndices(position_indices, orientation_indices, orientation_offset);
 
   fuse_core::VectorXd pose_mean_partial(position_indices.size() + orientation_indices.size());
   fuse_core::MatrixXd pose_covariance_partial(pose_mean_partial.rows(), pose_mean_partial.rows());
 
-  fuse_rl_common::populatePartialMeasurement(
+  fm_common::populatePartialMeasurement(
     pose_mean,
     pose_covariance,
     merged_indices,
@@ -223,12 +223,12 @@ TEST(TestSuite, populatePartialMeasurementEmptyPositionEmptyOrientation)
 
   const size_t orientation_offset = 2;
 
-  const auto merged_indices = fuse_rl_common::mergeIndices(position_indices, orientation_indices, orientation_offset);
+  const auto merged_indices = fm_common::mergeIndices(position_indices, orientation_indices, orientation_offset);
 
   fuse_core::VectorXd pose_mean_partial(position_indices.size() + orientation_indices.size());
   fuse_core::MatrixXd pose_covariance_partial(pose_mean_partial.rows(), pose_mean_partial.rows());
 
-  fuse_rl_common::populatePartialMeasurement(
+  fm_common::populatePartialMeasurement(
     pose_mean,
     pose_covariance,
     merged_indices,
