+++ fuse_models/test/test_unicycle_2d_ignition.cpp
@@ -35,9 +35,9 @@
 #include <fuse_core/eigen.h>
 #include <fuse_core/eigen_gtest.h>
 #include <fuse_core/transaction.h>
-#include <fuse_rl/unicycle_2d/ignition.h>
-#include <fuse_rl/SetPose.h>
-#include <fuse_rl/SetPoseDeprecated.h>
+#include <fuse_models/unicycle_2d/ignition.h>
+#include <fuse_models/SetPose.h>
+#include <fuse_models/SetPoseDeprecated.h>
 #include <ros/ros.h>
 
 #include <gtest/gtest.h>
@@ -99,7 +99,7 @@ TEST(Unicycle2DIgnition, InitialTransaction)
   auto callback_future = callback_promise.get_future();
 
   // Create an ignition sensor and register the callback
-  fuse_rl::unicycle_2d::Ignition ignition_sensor;
+  fuse_models::unicycle_2d::Ignition ignition_sensor;
   ignition_sensor.initialize("ignition_sensor", &transactionCallback);
   ignition_sensor.start();
 
@@ -175,7 +175,7 @@ TEST(Unicycle2DIgnition, SkipInitialTransaction)
   auto callback_future = callback_promise.get_future();
 
   // Create an ignition sensor and register the callback
-  fuse_rl::unicycle_2d::Ignition ignition_sensor;
+  fuse_models::unicycle_2d::Ignition ignition_sensor;
   ignition_sensor.initialize("ignition_sensor", &transactionCallback);
   ignition_sensor.start();
 
@@ -200,12 +200,12 @@ TEST(Unicycle2DIgnition, SetPoseService)
   auto callback_future = callback_promise.get_future();
 
   // Create an ignition sensor and register the callback
-  fuse_rl::unicycle_2d::Ignition ignition_sensor;
+  fuse_models::unicycle_2d::Ignition ignition_sensor;
   ignition_sensor.initialize("ignition_sensor", &transactionCallback);
   ignition_sensor.start();
 
   // Call the SetPose service
-  fuse_rl::SetPose srv;
+  fuse_models::SetPose srv;
   srv.request.pose.header.stamp = ros::Time(12, 345678910);
   srv.request.pose.pose.pose.position.x = 1.0;
   srv.request.pose.pose.pose.position.y = 2.0;
@@ -292,12 +292,12 @@ TEST(Unicycle2DIgnition, SetPoseDeprecatedService)
   auto callback_future = callback_promise.get_future();
 
   // Create an ignition sensor and register the callback
-  fuse_rl::unicycle_2d::Ignition ignition_sensor;
+  fuse_models::unicycle_2d::Ignition ignition_sensor;
   ignition_sensor.initialize("ignition_sensor", &transactionCallback);
   ignition_sensor.start();
 
   // Call the SetPose service
-  fuse_rl::SetPoseDeprecated srv;
+  fuse_models::SetPoseDeprecated srv;
   srv.request.pose.header.stamp = ros::Time(12, 345678910);
   srv.request.pose.pose.pose.position.x = 1.0;
   srv.request.pose.pose.pose.position.y = 2.0;
